Add launches model, controller and router
The launches model, controller, and routes are created in this update. In addition to the creation of the launches model, which holds launch object data, the 'getAllLaunches' method was implemented in the controller to provide all launches in json format. Furthermore, the 'httpGetLaunches' function was updated to fetch the launches data on the client-side. The server-side app file also now uses the 'launchesRouter'.